requirements.txt
aiosmtpd==1.4.6
alembic==1.13.2
atpublic==4.1.0
attrs==23.2.0
Babel==2.15.0
blinker==1.8.2
cachetools==4.2.4
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
dnspython==2.6.1
elastic-transport==8.13.1
elasticsearch==8.14.0
email_validator==2.2.0
Flask==3.0.3
flask-babel==4.0.0
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-Migrate==4.0.7
Flask-Moment==1.0.6
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
google-api-core==1.34.1
google-auth==1.35.0
google-cloud-core==1.7.3
google-cloud-translate==2.0.1
googleapis-common-protos==1.63.2
greenlet==3.0.3
grpcio==1.64.1
grpcio-status==1.48.2
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
lingua-language-detector==2.0.2
Mako==1.3.5
MarkupSafe==2.1.5
packaging==24.1
proto-plus==1.24.0
protobuf==3.20.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
PyJWT==2.8.0
python-dotenv==1.0.1
pytz==2024.1
redis==5.0.7
requests==2.32.3
rq==1.16.2
rsa==4.9
setuptools==70.3.0
six==1.16.0
SQLAlchemy==2.0.31
typing_extensions==4.12.2
urllib3==2.2.2
Werkzeug==3.0.3
WTForms==3.1.2
